Rename cli.rb to server.rb and remove unnecessary namespace
