+++ exe/ruby-lsp
@@ -20,4 +20,4 @@ end
 
 require_relative "../lib/ruby_lsp/internal"
 
-RubyLsp::Cli.start
+RubyLsp.start
