+++ lib/ruby_lsp/handler.rb
@@ -14,6 +14,13 @@ module RubyLsp
     extend T::Sig
     VOID = T.let(Object.new.freeze, Object)
 
+    sig { params(blk: T.proc.bind(Handler).params(arg0: T.untyped).void).void }
+    def self.start(&blk)
+      handler = new
+      handler.instance_exec(&blk)
+      handler.start
+    end
+
     sig { returns(Store) }
     attr_reader :store
 
@@ -29,11 +36,6 @@ module RubyLsp
     def start
       $stderr.puts "Starting Ruby LSP..."
       @reader.read { |request| handle(request) }
-    end
-
-    sig { params(blk: T.proc.bind(Handler).params(arg0: T.untyped).void).void }
-    def config(&blk)
-      instance_exec(&blk)
     end
 
     private
